Restructured the project into Presentation, Domain and Data packages
