+++ app/src/main/AndroidManifest.xml
@@ -15,13 +15,13 @@
         android:theme="@style/Theme.EcoMonitor"
         tools:targetApi="31">
         <activity
-            android:name=".MainMenuActivity"
+            android:name=".presentation.view.MainMenuActivity"
             android:exported="false" />
         <activity
-            android:name=".SignUpActivity"
+            android:name=".presentation.view.SignUpActivity"
             android:exported="false" />
         <activity
-            android:name=".SignInActivity"
+            android:name=".presentation.view.SignInActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
