+++ app/src/main/java/com/example/ecomonitor/presentation/view/MainMenuActivity.kt
@@ -1,7 +1,8 @@
-package com.example.ecomonitor
+package com.example.ecomonitor.presentation.view
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
+import com.example.ecomonitor.R
 
 class MainMenuActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
