+++ app/src/main/res/layout/activity_main_menu.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainMenuActivity">
+    tools:context=".presentation.view.MainMenuActivity">
 
     <TextView
         android:id="@+id/textView8"
