+++ app/src/main/res/layout/activity_sign_in.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:elevation="2dp"
     android:textAlignment="textEnd"
-    tools:context=".SignInActivity"
+    tools:context=".presentation.view.SignInActivity"
     tools:layout_editor_absoluteX="-66dp"
     tools:layout_editor_absoluteY="0dp">
